✨ Refactor weather UI structure and add store badges for Web
This commit reorganizes the weather module's UI components into a more structured directory hierarchy and introduces app store badges to the persistent footer when running on the web.

### UI & Layout
- **`lib/about/about_page.dart`**: Added `StoreBadge` widgets (Google Play and App Store) to the `persistentFooterButtons` when the app is running on Web.
- **`lib/weather/ui/page/weather_page_default_layout.dart`**: Added store badges to the footer for web users on smaller screens.
- **`lib/support/support_page.dart`**: Adjusted padding and added a `SizedBox` for better spacing at the top of the list view.

### Refactoring & File Reorganization
- Moved weather pages to `lib/weather/ui/page/`:
    - `weather_page.dart`
    - `weather_page_default_layout.dart`
    - `weather_page_extra_small_layout.dart`
- Moved weather-specific widgets to `lib/weather/ui/widgets/`:
    - `outfit_widget.dart`
    - `weather_loading_widget.dart`
    - `weather_icon.dart`
    - `outfit_image_widget.dart`
    - `text_recommendation_widget.dart`
    - `loading_outfit_text_widget.dart`
- Relocated `weather.dart` (export file) from `lib/weather/ui/` to `lib/weather/` and updated its export paths.
- Updated all internal imports and test references to reflect the new directory structure.
