+++ lib/weather/ui/widgets/outfit_image_widget.dart
